+++ core/src/test/java/com/github/skjolber/packing/packer/laff/LargestAreaFitFirstPackagerWeightConstraintTest.java
@@ -210,8 +210,8 @@ public class LargestAreaFitFirstPackagerWeightConstraintTest extends AbstractPac
 			List<Placement> placements = result.getContainers().get(0).getStack().getPlacements();
 			Placement bot = placementAt(placements, 0);
 			Placement top = placementAt(placements, 1);
-			assertThat(bot.getLoadWeight()).isEqualTo(top.getWeight());
-			assertThat(top.getLoadWeight()).isEqualTo(0L);
+			assertThat(bot.getLoadWeight()).isEqualTo((double) top.getWeight());
+			assertThat(top.getLoadWeight()).isEqualTo(0.0);
 		} finally {
 			packager.close();
 		}
